refactor: separate user scoped config from global config
