+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/configuration/UserConfigRepository.kt
@@ -4,39 +4,23 @@ import com.wire.kalium.logic.StorageFailure
 import com.wire.kalium.logic.functional.Either
 import com.wire.kalium.logic.functional.map
 import com.wire.kalium.logic.wrapStorageRequest
-import com.wire.kalium.persistence.client.UserConfigStorage
+import com.wire.kalium.persistence.config.UserConfigStorage
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
 interface UserConfigRepository {
-    fun persistEnableLogging(enabled: Boolean): Either<StorageFailure, Unit>
-    fun isLoggingEnabled(): Either<StorageFailure, Boolean>
-    fun isPersistentWebSocketConnectionEnabledFlow(): Flow<Either<StorageFailure, Boolean>>
-    fun persistPersistentWebSocketConnectionStatus(status: Boolean): Either<StorageFailure, Unit>
     fun setFileSharingStatus(status: Boolean, isStatusChanged: Boolean?): Either<StorageFailure, Unit>
     fun isFileSharingEnabled(): Either<StorageFailure, FileSharingStatus>
     fun isFileSharingEnabledFlow(): Flow<Either<StorageFailure, FileSharingStatus>>
-    fun isMLSEnabled(): Either<StorageFailure, Boolean>
-    fun setMLSEnabled(enabled: Boolean): Either<StorageFailure, Unit>
     fun setClassifiedDomainsStatus(enabled: Boolean, domains: List<String>): Either<StorageFailure, Unit>
     fun getClassifiedDomainsStatus(): Flow<Either<StorageFailure, ClassifiedDomainsStatus>>
+    fun isMLSEnabled(): Either<StorageFailure, Boolean>
+    fun setMLSEnabled(enabled: Boolean): Either<StorageFailure, Unit>
 }
 
 class UserConfigDataSource(
     private val userConfigStorage: UserConfigStorage
 ) : UserConfigRepository {
-
-    override fun persistEnableLogging(enabled: Boolean): Either<StorageFailure, Unit> =
-        wrapStorageRequest { userConfigStorage.enableLogging(enabled) }
-
-    override fun isLoggingEnabled(): Either<StorageFailure, Boolean> =
-        wrapStorageRequest { userConfigStorage.isLoggingEnables() }
-
-    override fun isPersistentWebSocketConnectionEnabledFlow(): Flow<Either<StorageFailure, Boolean>> =
-        userConfigStorage.isPersistentWebSocketConnectionEnabledFlow().wrapStorageRequest()
-
-    override fun persistPersistentWebSocketConnectionStatus(status: Boolean): Either<StorageFailure, Unit> =
-        wrapStorageRequest { userConfigStorage.persistPersistentWebSocketConnectionStatus(status) }
 
     override fun setFileSharingStatus(status: Boolean, isStatusChanged: Boolean?): Either<StorageFailure, Unit> =
         wrapStorageRequest { userConfigStorage.persistFileSharingStatus(status, isStatusChanged) }
@@ -58,12 +42,6 @@ class UserConfigDataSource(
                 }
             }
 
-    override fun isMLSEnabled(): Either<StorageFailure, Boolean> =
-        wrapStorageRequest { userConfigStorage.isMLSEnabled() }
-
-    override fun setMLSEnabled(enabled: Boolean): Either<StorageFailure, Unit> =
-        wrapStorageRequest { userConfigStorage.enableMLS(enabled) }
-
     override fun setClassifiedDomainsStatus(enabled: Boolean, domains: List<String>) =
         wrapStorageRequest { userConfigStorage.persistClassifiedDomainsStatus(enabled, domains) }
 
@@ -73,4 +51,10 @@ class UserConfigDataSource(
                 ClassifiedDomainsStatus(classifiedDomain.status, classifiedDomain.trustedDomains)
             }
         }
+
+    override fun isMLSEnabled(): Either<StorageFailure, Boolean> =
+        wrapStorageRequest { userConfigStorage.isMLSEnabled() }
+
+    override fun setMLSEnabled(enabled: Boolean): Either<StorageFailure, Unit> =
+        wrapStorageRequest { userConfigStorage.enableMLS(enabled) }
 }
