+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/user/webSocketStatus/PersistPersistentWebSocketConnectionStatusUseCase.kt
@@ -2,7 +2,7 @@ package com.wire.kalium.logic.feature.user.webSocketStatus
 
 import com.wire.kalium.logger.KaliumLogger.Companion.ApplicationFlow.LOCAL_STORAGE
 import com.wire.kalium.logic.StorageFailure
-import com.wire.kalium.logic.configuration.UserConfigRepository
+import com.wire.kalium.logic.configuration.GlobalConfigRepository
 import com.wire.kalium.logic.functional.fold
 import com.wire.kalium.logic.kaliumLogger
 
@@ -11,10 +11,10 @@ interface PersistPersistentWebSocketConnectionStatusUseCase {
 }
 
 internal class PersistPersistentWebSocketConnectionStatusUseCaseImpl(
-    private val userConfigRepository: UserConfigRepository
+    private val globalConfigRepository: GlobalConfigRepository
 ) : PersistPersistentWebSocketConnectionStatusUseCase {
     override operator fun invoke(enabled: Boolean): Unit =
-        userConfigRepository.persistPersistentWebSocketConnectionStatus(enabled).let {
+        globalConfigRepository.persistPersistentWebSocketConnectionStatus(enabled).let {
             it.fold({ storageFailure ->
                 when (storageFailure) {
                     is StorageFailure.DataNotFound ->
