+++ persistence/src/androidMain/kotlin/com/wire/kalium/persistence/kmmSettings/GlobalPrefProvider.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import com.wire.kalium.persistence.client.AuthTokenStorage
 import com.wire.kalium.persistence.client.TokenStorage
 import com.wire.kalium.persistence.client.TokenStorageImpl
-import com.wire.kalium.persistence.client.UserConfigStorage
-import com.wire.kalium.persistence.client.UserConfigStorageImpl
+import com.wire.kalium.persistence.config.GlobalAppConfigStorage
+import com.wire.kalium.persistence.config.GlobalAppConfigStorageImpl
 import com.wire.kalium.persistence.dbPassphrase.PassphraseStorage
 import com.wire.kalium.persistence.dbPassphrase.PassphraseStorageImpl
 
@@ -25,5 +25,6 @@ actual class GlobalPrefProvider(context: Context, shouldEncryptData: Boolean = t
         get() = PassphraseStorageImpl(encryptedSettingsHolder)
     actual val tokenStorage: TokenStorage
         get() = TokenStorageImpl(encryptedSettingsHolder)
-    actual val userConfigStorage: UserConfigStorage = UserConfigStorageImpl(encryptedSettingsHolder)
+    actual val globalAppConfigStorage: GlobalAppConfigStorage = GlobalAppConfigStorageImpl(encryptedSettingsHolder)
+
 }
