+++ persistence/src/androidMain/kotlin/com/wire/kalium/persistence/kmmSettings/UserPrefProvider.kt
@@ -3,6 +3,8 @@ package com.wire.kalium.persistence.kmmSettings
 import android.content.Context
 import com.wire.kalium.persistence.client.LastRetrievedNotificationEventStorage
 import com.wire.kalium.persistence.client.LastRetrievedNotificationEventStorageImpl
+import com.wire.kalium.persistence.config.UserConfigStorage
+import com.wire.kalium.persistence.config.UserConfigStorageImpl
 import com.wire.kalium.persistence.dao.UserIDEntity
 
 actual class UserPrefProvider(
@@ -20,6 +22,8 @@ actual class UserPrefProvider(
 
     actual val lastRetrievedNotificationEventStorage: LastRetrievedNotificationEventStorage
         get() = LastRetrievedNotificationEventStorageImpl(encryptedSettingsHolder)
+
+    actual val userConfigStorage: UserConfigStorage = UserConfigStorageImpl(encryptedSettingsHolder)
 
     actual fun clear() {
         encryptedSettingsHolder.nuke()
