+++ persistence/src/iosX64Main/kotlin/com/wire/kalium/persistence/kmmSettings/GlobalPrefProvider.kt
@@ -2,7 +2,7 @@ package com.wire.kalium.persistence.kmmSettings
 
 import com.wire.kalium.persistence.client.AuthTokenStorage
 import com.wire.kalium.persistence.client.TokenStorage
-import com.wire.kalium.persistence.client.UserConfigStorage
+import com.wire.kalium.persistence.config.GlobalAppConfigStorage
 import com.wire.kalium.persistence.dbPassphrase.PassphraseStorage
 
 actual class GlobalPrefProvider {
@@ -12,8 +12,5 @@ actual class GlobalPrefProvider {
         get() = TODO("Not yet implemented")
     actual val tokenStorage: TokenStorage
         get() = TODO("Not yet implemented")
-
-    @Deprecated("must be moved to user specific storage")
-    actual val userConfigStorage: UserConfigStorage
-        get() = TODO("Not yet implemented")
+    actual val globalAppConfigStorage: GlobalAppConfigStorage = TODO("Not yet implemented")
 }
