+++ persistence/src/iosX64Main/kotlin/com/wire/kalium/persistence/kmmSettings/UserPrefProvider.kt
@@ -1,6 +1,7 @@
 package com.wire.kalium.persistence.kmmSettings
 
 import com.wire.kalium.persistence.client.LastRetrievedNotificationEventStorage
+import com.wire.kalium.persistence.config.UserConfigStorage
 
 actual class UserPrefProvider {
 
@@ -12,5 +13,8 @@ actual class UserPrefProvider {
     actual fun clear() {
         TODO("Not yet implemented")
     }
+
+    actual val userConfigStorage: UserConfigStorage
+        get() = TODO("Not yet implemented")
 
 }
