+++ persistence/src/jsMain/kotlin/com/wire/kalium/persistence/kmmSettings/UserPrefProvider.kt
@@ -1,6 +1,7 @@
 package com.wire.kalium.persistence.kmmSettings
 
 import com.wire.kalium.persistence.client.LastRetrievedNotificationEventStorage
+import com.wire.kalium.persistence.config.UserConfigStorage
 
 actual class UserPrefProvider {
     actual val lastRetrievedNotificationEventStorage: LastRetrievedNotificationEventStorage
@@ -9,4 +10,7 @@ actual class UserPrefProvider {
     actual fun clear() {
         TODO("Not yet implemented")
     }
+
+    actual val userConfigStorage: UserConfigStorage
+        get() = TODO("Not yet implemented")
 }
