+++ persistence/src/jvmMain/kotlin/com/wire/kalium/persistence/kmmSettings/GlobalPrefProvider.kt
@@ -3,8 +3,8 @@ package com.wire.kalium.persistence.kmmSettings
 import com.wire.kalium.persistence.client.AuthTokenStorage
 import com.wire.kalium.persistence.client.TokenStorage
 import com.wire.kalium.persistence.client.TokenStorageImpl
-import com.wire.kalium.persistence.client.UserConfigStorage
-import com.wire.kalium.persistence.client.UserConfigStorageImpl
+import com.wire.kalium.persistence.config.GlobalAppConfigStorage
+import com.wire.kalium.persistence.config.GlobalAppConfigStorageImpl
 import com.wire.kalium.persistence.dbPassphrase.PassphraseStorage
 import com.wire.kalium.persistence.dbPassphrase.PassphraseStorageImpl
 
@@ -21,5 +21,5 @@ actual class GlobalPrefProvider(
         get() = PassphraseStorageImpl(kaliumPref)
     actual val tokenStorage: TokenStorage
         get() = TokenStorageImpl(kaliumPref)
-    actual val userConfigStorage: UserConfigStorage = UserConfigStorageImpl(kaliumPref)
+    actual val globalAppConfigStorage: GlobalAppConfigStorage = GlobalAppConfigStorageImpl(kaliumPref)
 }
