+++ persistence/src/jvmMain/kotlin/com/wire/kalium/persistence/kmmSettings/UserPrefProvider.kt
@@ -2,6 +2,8 @@ package com.wire.kalium.persistence.kmmSettings
 
 import com.wire.kalium.persistence.client.LastRetrievedNotificationEventStorage
 import com.wire.kalium.persistence.client.LastRetrievedNotificationEventStorageImpl
+import com.wire.kalium.persistence.config.UserConfigStorage
+import com.wire.kalium.persistence.config.UserConfigStorageImpl
 import com.wire.kalium.persistence.dao.UserIDEntity
 
 actual class UserPrefProvider(
@@ -17,8 +19,10 @@ actual class UserPrefProvider(
 
     actual val lastRetrievedNotificationEventStorage: LastRetrievedNotificationEventStorage
         get() = LastRetrievedNotificationEventStorageImpl(kaliumPref)
+    actual val userConfigStorage: UserConfigStorage = UserConfigStorageImpl(kaliumPref)
 
     actual fun clear() {
         kaliumPref.nuke()
     }
+
 }
